requirements.txt
openai
python-docx
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
language-tool-python
python-dotenv
setuptools
psutil